2011-09-02  chl  <chl@tuxfamily.org>
	* configure.ac:
	- update from openssh's upstream repository

	* openbsd-compat/Makefile.am:
	* openbsd-compat/daemon.c:
	- add daemon()

	* src/libexec/mail.local/mail.local.c:
	- add missing header

+++ src/libexec/mail.local/mail.local.c
@@ -48,6 +48,8 @@
 #include "pathnames.h"
 #include "mail.local.h"
 
+#include "openbsd-compat.h"
+
 int
 main(int argc, char *argv[])
 {
@@ -254,7 +256,8 @@ retry:
 	}
 
 	curoff = lseek(mbfd, 0, SEEK_END);
-	(void)snprintf(biffmsg, sizeof biffmsg, "%s@%qd\n", name, curoff);
+	(void)snprintf(biffmsg, sizeof biffmsg, "%s@%qd\n", name,
+		       (long long int) curoff);
 	if (lseek(fd, 0, SEEK_SET) == (off_t)-1) {
 		merr(NOTFATAL, "temporary file: %s", strerror(errno));
 		goto bad;
